+++ src/test/java/com/shyoon/wms/outbound/feature/RegisterPackagingMaterialTest.java
@@ -3,12 +3,6 @@ package com.shyoon.wms.outbound.feature;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.util.Assert;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -55,231 +49,4 @@ class RegisterPackagingMaterialTest {
         assertThat(packagingMaterialRepository.findAll()).hasSize(1);
     }
 
-    private class RegisterPackagingMaterial {
-
-        private final PackagingMaterialRepository packagingMaterialRepository;
-
-        private RegisterPackagingMaterial(PackagingMaterialRepository packagingMaterialRepository) {
-            this.packagingMaterialRepository = packagingMaterialRepository;
-        }
-
-        public void request(final Request request) {
-            final PackagingMaterial packagingMaterial = request.toDomain();
-
-            packagingMaterialRepository.save(packagingMaterial);
-        }
-
-        public record Request(
-                String name,
-                String code,
-                Long innerWidthInMillimeters,
-                Long innerHeightInMillimeters,
-                Long innerLengthInMillimeters,
-                Long outerWidthInMillimeters,
-                Long outerHeightInMillimeters,
-                Long outerLengthInMillimeters,
-                Long weightInGrams,
-                Long maxWeightInGrams,
-                MaterialType materialType) {
-
-            public Request {
-                Assert.hasText(name, "포장재 이름은 필수입니다.");
-                Assert.hasText(code, "포장재 코드는 필수입니다.");
-                Assert.notNull(innerWidthInMillimeters, "내부 폭은 필수입니다.");
-                if (innerWidthInMillimeters < 1) {
-                    throw new IllegalArgumentException("내부 폭은 1mm 이상이어야 합니다.");
-                }
-                Assert.notNull(innerHeightInMillimeters, "내부 높이는 필수입니다.");
-                if (innerHeightInMillimeters < 1) {
-                    throw new IllegalArgumentException("내부 높이는 1mm 이상이어야 합니다.");
-                }
-                Assert.notNull(innerLengthInMillimeters, "내부 길이는 필수입니다.");
-                if (innerLengthInMillimeters < 1) {
-                    throw new IllegalArgumentException("내부 길이는 1mm 이상이어야 합니다.");
-                }
-                Assert.notNull(outerWidthInMillimeters, "외부 폭은 필수입니다.");
-                if (outerWidthInMillimeters < 1) {
-                    throw new IllegalArgumentException("외부 폭은 1mm 이상이어야 합니다.");
-                }
-                Assert.notNull(outerHeightInMillimeters, "외부 높이는 필수입니다.");
-                if (outerHeightInMillimeters < 1) {
-                    throw new IllegalArgumentException("외부 높이는 1mm 이상이어야 합니다.");
-                }
-                Assert.notNull(outerLengthInMillimeters, "외부 길이는 필수입니다.");
-                if (outerLengthInMillimeters < 1) {
-                    throw new IllegalArgumentException("외부 길이는 1mm 이상이어야 합니다.");
-                }
-                Assert.notNull(weightInGrams, "무게는 필수입니다.");
-                if (weightInGrams < 1) {
-                    throw new IllegalArgumentException("무게는 1g 이상이어야 합니다.");
-                }
-                Assert.notNull(maxWeightInGrams, "최대 무게는 필수입니다.");
-                if (maxWeightInGrams < 1) {
-                    throw new IllegalArgumentException("최대 무게는 1g 이상이어야 합니다.");
-                }
-                Assert.notNull(materialType, "포장재 종류는 필수입니다.");
-            }
-
-            public PackagingMaterial toDomain() {
-                return new PackagingMaterial(
-                        name,
-                        code,
-                        new PackagingMaterialDemension(
-                                innerWidthInMillimeters,
-                                innerHeightInMillimeters,
-                                innerLengthInMillimeters,
-                                outerWidthInMillimeters,
-                                outerHeightInMillimeters,
-                                outerLengthInMillimeters
-                        ),
-                        weightInGrams,
-                        maxWeightInGrams,
-                        materialType
-                );
-            }
-
-        }
-
-
-    }
-
-    enum MaterialType {
-        CORRUGATED_BOX("완충재가 있는 골판지 상자");
-
-        private final String description;
-
-        MaterialType(String description) {
-            this.description = description;
-        }
-    }
-
-    private static class PackagingMaterial {
-        private Long packagingMaterialNo;
-        private final String name;
-        private final String code;
-        private final PackagingMaterialDemension packagingMaterialDemension;
-        private final Long weightInGrams;
-        private final Long maxWeightInGrams;
-        private final MaterialType materialType;
-
-        public PackagingMaterial(
-                final String name,
-                final String code,
-                final PackagingMaterialDemension packagingMaterialDemension,
-                final Long weightInGrams,
-                final Long maxWeightInGrams,
-                final MaterialType materialType) {
-            validateConstructor(name, code, packagingMaterialDemension, weightInGrams, maxWeightInGrams, materialType);
-
-            this.name = name;
-            this.code = code;
-            this.packagingMaterialDemension = packagingMaterialDemension;
-            this.weightInGrams = weightInGrams;
-            this.maxWeightInGrams = maxWeightInGrams;
-            this.materialType = materialType;
-        }
-
-        private void validateConstructor(
-                final String name,
-                final String code,
-                final PackagingMaterialDemension packagingMaterialDemension,
-                final Long weightInGrams,
-                final Long maxWeightInGrams,
-                final MaterialType materialType) {
-            Assert.hasText(name, "포장재 이름은 필수입니다.");
-            Assert.hasText(code, "포장재 코드는 필수입니다.");
-            Assert.notNull(packagingMaterialDemension, "포장재 치수는 필수입니다.");
-            Assert.notNull(weightInGrams, "무게는 필수입니다.");
-            Assert.notNull(maxWeightInGrams, "최대 무게는 필수입니다.");
-            Assert.notNull(materialType, "포장재 종류는 필수입니다.");
-        }
-
-        public void assignNo(Long packagingMaterialNo) {
-            this.packagingMaterialNo = packagingMaterialNo;
-        }
-
-        public Long getPackagingMaterialNo() {
-            return packagingMaterialNo;
-        }
-    }
-
-    private static class PackagingMaterialDemension {
-        private final Long innerWidthInMillimeters;
-        private final Long innerHeightInMillimeters;
-        private final Long innerLengthInMillimeters;
-        private final Long outerWidthInMillimeters;
-        private final Long outerHeightInMillimeters;
-        private final Long outerLengthInMillimeters;
-
-        public PackagingMaterialDemension(
-                final Long innerWidthInMillimeters,
-                final Long innerHeightInMillimeters,
-                final Long innerLengthInMillimeters,
-                final Long outerWidthInMillimeters,
-                final Long outerHeightInMillimeters,
-                final Long outerLengthInMillimeters) {
-            validateConstructor(
-                    innerWidthInMillimeters,
-                    innerHeightInMillimeters,
-                    innerLengthInMillimeters,
-                    outerWidthInMillimeters,
-                    outerHeightInMillimeters,
-                    outerLengthInMillimeters);
-
-            this.innerWidthInMillimeters = innerWidthInMillimeters;
-            this.innerHeightInMillimeters = innerHeightInMillimeters;
-            this.innerLengthInMillimeters = innerLengthInMillimeters;
-            this.outerWidthInMillimeters = outerWidthInMillimeters;
-            this.outerHeightInMillimeters = outerHeightInMillimeters;
-            this.outerLengthInMillimeters = outerLengthInMillimeters;
-        }
-
-        private void validateConstructor(
-                final Long innerWidthInMillimeters,
-                final Long innerHeightInMillimeters,
-                final Long innerLengthInMillimeters,
-                final Long outerWidthInMillimeters,
-                final Long outerHeightInMillimeters,
-                final Long outerLengthInMillimeters) {
-            Assert.notNull(innerWidthInMillimeters, "내부 폭은 필수입니다.");
-            if (innerWidthInMillimeters < 1) {
-                throw new IllegalArgumentException("내부 폭은 1mm 이상이어야 합니다.");
-            }
-            Assert.notNull(innerHeightInMillimeters, "내부 높이는 필수입니다.");
-            if (innerHeightInMillimeters < 1) {
-                throw new IllegalArgumentException("내부 높이는 1mm 이상이어야 합니다.");
-            }
-            Assert.notNull(innerLengthInMillimeters, "내부 길이는 필수입니다.");
-            if (innerLengthInMillimeters < 1) {
-                throw new IllegalArgumentException("내부 길이는 1mm 이상이어야 합니다.");
-            }
-            Assert.notNull(outerWidthInMillimeters, "외부 폭은 필수입니다.");
-            if (outerWidthInMillimeters < 1) {
-                throw new IllegalArgumentException("외부 폭은 1mm 이상이어야 합니다.");
-            }
-            Assert.notNull(outerHeightInMillimeters, "외부 높이는 필수입니다.");
-            if (outerHeightInMillimeters < 1) {
-                throw new IllegalArgumentException("외부 높이는 1mm 이상이어야 합니다.");
-            }
-            Assert.notNull(outerLengthInMillimeters, "외부 길이는 필수입니다.");
-            if (outerLengthInMillimeters < 1) {
-                throw new IllegalArgumentException("외부 길이는 1mm 이상이어야 합니다.");
-            }
-        }
-    }
-
-    private class PackagingMaterialRepository {
-
-        private final Map<Long, PackagingMaterial> packagingMaterialMap = new HashMap<>();
-        private Long sequence = 1L;
-
-        public void save(PackagingMaterial packagingMaterial) {
-            packagingMaterial.assignNo(sequence++);
-            packagingMaterialMap.put(packagingMaterial.getPackagingMaterialNo(), packagingMaterial);
-        }
-
-        public List<PackagingMaterial> findAll() {
-            return new ArrayList<>(packagingMaterialMap.values());
-        }
-    }
 }
